CV-ATS-Checker/requirements.txt
annotated-types
anyio
blis
catalogue
certifi
cffi
charset-normalizer
click
cloudpathlib
confection
cryptography
cymem
deprecation
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl
et_xmlfile
fastapi
h11
h2
hpack
httpcore
httpx
hyperframe
idna
Jinja2
joblib
langcodes
language_data
lxml
marisa-trie
markdown-it-py
MarkupSafe
mdurl
murmurhash
nltk
numpy
openpyxl
packaging
pandas
pdf2image
pdfminer.six
pillow
postgrest
preshed
pycparser
pydantic
pydantic_core
Pygments
PyJWT
PyMuPDF
pytesseract
python-dateutil
python-docx
python-dotenv
python-multipart
pytz
realtime
regex
requests
rich
setuptools
shellingham
six
smart_open
sniffio
spacy
spacy-legacy
spacy-loggers
srsly
starlette
storage3
StrEnum
supabase
supabase_auth
supabase_functions
thinc
tqdm
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
wasabi
weasel
websockets
wrapt